requirements.txt
apache-airflow[amazon]
pyspark~=3.5.3
apache-airflow
boto3~=1.35.44
